Search component added
