+++ src/common/Header/styled.js
@@ -1,7 +1,6 @@
 import styled from "styled-components";
 import { Link, NavLink } from "react-router-dom";
 import { ReactComponent as VideoIcon } from "./images/VideoIcon.svg";
-import { ReactComponent as SearchIcon } from "./images/SearchIcon.svg";
 const mobileBp = ({ theme }) => theme.breakpoints.mobileMax;
 
 export const StyledHeader = styled.header`
@@ -104,55 +103,5 @@ export const StyledNavLink = styled(NavLink)`
   @media (max-width: ${mobileBp}px) {
     font-size: 12px;
     padding: 8px 12px;
-  }
-`;
-
-export const SearchBar = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-export const StyledSearchIcon = styled(SearchIcon)`
-  @media (max-width: ${mobileBp}px) {
-    width: 16px;
-  }
-`;
-
-export const SearchBarIcon = styled.div`
-  color: ${({ theme }) => theme.colors.waterloo};
-  background-color: ${({ theme }) => theme.colors.white};
-  height: 44px;
-  width: 50px;
-  border-radius: 32px 0 0 33px;
-  display: flex;
-  align-items: center;
-  justify-content: right;
-
-  @media (max-width: ${mobileBp}px) {
-    height: 44px;
-    width: 35px;
-  }
-`;
-
-export const SearchBarInput = styled.input`
-  height: 44px;
-  width: 100%;
-  padding: 19px;
-  border: 1px solid ${({ theme }) => theme.colors.mystic};
-  border-left: none;
-  border-radius: 0 33px 33px 0;
-
-  &::placeholder {
-    font-size: 16px;
-  }
-
-  &:focus {
-    outline: none;
-  }
-
-  @media (max-width: ${mobileBp}px) {
-    height: 44px;
-    width: 100%;
-    padding: 10px;
   }
 `;
