+++ src/features/Search/index.js
@@ -1,0 +1,14 @@
+import { SearchBar, SearchBarIcon, StyledSearchIcon, SearchBarInput } from "./styled"
+
+export const Search = () => {
+    return (
+        <SearchBar>
+            <SearchBarIcon>
+                <StyledSearchIcon />
+            </SearchBarIcon>
+            <SearchBarInput
+                placeholder="Search for movies/people..."
+            />
+        </SearchBar>
+    )
+}
